Reduce number of enums called BlendMode
